Add func to get player start from level array
This will make it easier to design future levels by placing the player
start directly in the array.

Have to remove the player start from the array during initialization
because it is only used for the player start. If left in, it will get in
the way when moving around.